test: split target selection and validation out of test_connect_cloud.py
Target inference and post-resolution option validation are not Connect
Cloud features: a deployment record may name a Posit Connect server,
shinyapps.io, or Connect Cloud, and the shinyapps.io option checks are not
Cloud-related at all. Both were tested only inside test_connect_cloud.py,
which was the only file in the suite asserting on any validation error
message -- so someone changing shinyapps.io handling or looking for
test_validation.py found nothing.

- tests/test_target_inference.py: TestRedeployTargetInference and
  TestInferenceIsWiredToTheCommands (61 tests).
- tests/test_validation.py: TestConnectOptionsAgainstAResolvedEntry and
  TestConnectOnlyDeployOptions (13 tests).
- tests/target_helpers.py: the store, click-context, and CLI-introspection
  fixtures the three files share.

A pure move: no assertion changed, only imports. The full suite reports the
same 1180 passed, 12 skipped, 815 subtests before and after, and the 357
tests in the old file are now 283 + 61 + 13.
